Prepare extension for Frontend mode Use ENUM for configuration constant Possibility to deactivate chatbot in BE
