+++ Classes/Controller/Backend/ChatbotController.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Ameos\Chatbot\Controller;
+namespace Ameos\Chatbot\Controller\Backend;
 
 use Ameos\Chatbot\Service\ChatbotService;
+use Ameos\Chatbot\Service\Prompt\BackendPromptService;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Core\Http\JsonResponse;
@@ -13,9 +14,12 @@ final class ChatbotController
 {
     /**
      * @param ChatbotService $chatbotService
+     * @param BackendPromptService $backendPromptService
      */
-    public function __construct(private readonly ChatbotService $chatbotService)
-    {
+    public function __construct(
+        private readonly ChatbotService $chatbotService,
+        private readonly BackendPromptService $backendPromptService
+    ) {
     }
 
     /**
@@ -29,7 +33,10 @@ final class ChatbotController
         $body = json_decode($request->getBody()->getContents(), true);
 
         return new JsonResponse([
-            'message' => $this->chatbotService->request($body['message'] ?? '')
+            'message' => $this->chatbotService->request(
+                $body['message'] ?? '',
+                $this->backendPromptService->getPrompt()
+            )
         ]);
     }
 }
