+++ Configuration/Backend/Routes.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Ameos\Chatbot\Controller\ChatbotController;
+use Ameos\Chatbot\Controller\Backend\ChatbotController;
 
 return [
     'ameos_chatbot_question' => [
