+++ ext_emconf.php
@@ -7,7 +7,7 @@ $EM_CONF['chatbot'] = [
     'author_company'   => 'Ameos',
     'author_email'     => 'typo3dev@ameos.com',
     'state'            => 'beta',
-    'version'          => '1.1.0',
+    'version'          => '1.1.1',
     'constraints'      => [
         'depends' => [
             'typo3' => '12.4.0-12.4.99',
